Expose utils to servers
