+++ src/lib/main/utils/index.ts
@@ -1,6 +1,3 @@
-export * from "./errors";
 export * from "./extensions";
-export * from "./hex-to-bech32";
-export * from "./lovelace";
-export * from "./strip-address-prefix";
 export * from "./wallets";
+export * from "@/lib/server";
