+++ src/lib/server/index.ts
@@ -1,11 +1,6 @@
-export const STORAGE_KEYS = {
-  connectedWallet: "weld_connected-wallet",
-  connectedChange: "weld_connected-change",
-  connectedStake: "weld_connected-stake",
-  connectedSolWallet: "weld_connected-sol-wallet",
-  connectedEthWallet: "weld_connected-eth-wallet",
-  connectedPolyWallet: "weld_connected-poly-wallet",
-  connectedBtcWallet: "weld_connected-btc-wallet",
-} as const;
-
-export type StorageKeysType = (typeof STORAGE_KEYS)[keyof typeof STORAGE_KEYS];
+export * from "./address";
+export * from "./currency";
+export * from "./errors";
+export * from "./hex-to-bech32";
+export * from "./storage";
+export * from "./strip-address-prefix";
